[CPDLP-3074] TS5 - refactor .schedule_for
