+++ app/models/course.rb
@@ -43,8 +43,8 @@ class Course < ApplicationRecord
     npq-early-years-leadership
   ].freeze
 
-  def schedule_for(cohort:)
-    course_group.schedule_for(cohort:)
+  def schedule_for(cohort: Cohort.current, schedule_date: Date.current)
+    course_group.schedule_for(cohort:, schedule_date:)
   end
 
   class << self
